+++ dis-app/src/main/java/it/akademija/models/UserDataInfo.java
@@ -8,14 +8,14 @@ public class UserDataInfo {
 	private Long personId;
 	private String address;
 	private String city;
-	private Integer phoneNum;
+	private Long phoneNum;
 	private String email;
 	private User user;
 
 	public UserDataInfo() {
 	}
 
-	public UserDataInfo(String name, String surename, Long personId, String address, String city, Integer phoneNum,
+	public UserDataInfo(String name, String surename, Long personId, String address, String city, Long phoneNum,
 			String email) {
 		this.name = name;
 		this.surename = surename;
@@ -27,7 +27,7 @@ public class UserDataInfo {
 	}
 
 	public UserDataInfo(long id, String name, String surename, Long personId, String address, String city,
-			Integer phoneNum, String email) {
+			Long phoneNum, String email) {
 		this.id = id;
 		this.name = name;
 		this.surename = surename;
@@ -38,8 +38,8 @@ public class UserDataInfo {
 		this.email = email;
 	}
 
-	public UserDataInfo(long id, String name, String surename, Long personId, String address, String city, int phoneNum,
-			String email, User user) {
+	public UserDataInfo(long id, String name, String surename, Long personId, String address, String city,
+			Long phoneNum, String email, User user) {
 		this.id = id;
 		this.name = name;
 		this.surename = surename;
@@ -99,11 +99,11 @@ public class UserDataInfo {
 		this.city = city;
 	}
 
-	public Integer getPhoneNum() {
+	public Long getPhoneNum() {
 		return phoneNum;
 	}
 
-	public void setPhoneNum(Integer phoneNum) {
+	public void setPhoneNum(Long phoneNum) {
 		this.phoneNum = phoneNum;
 	}
